Replace SIGHUP with socket
