+++ tests/test_sockets.py
@@ -5,13 +5,15 @@
 import io
 import unittest
 import urllib.error
-from controlsocket import Client, APIError, ConnectionError
+from controlsocket import ControlSocketClient
+from configchangesocket import ConfigChangeSocketClient
+from unixsocket import APIError, ConnectionError
 
 
 class TestClass(unittest.TestCase):
     def test_add_metrics_user_success(self):
         mock_opener = MockOpener(self)
-        control_socket = Client('fake_socket_path', opener=mock_opener)
+        control_socket = ControlSocketClient('fake_socket_path', opener=mock_opener)
 
         mock_opener.expect(
             url='http://localhost/metrics-users',
@@ -27,7 +29,7 @@ class TestClass(unittest.TestCase):
 
     def test_add_metrics_user_fail(self):
         mock_opener = MockOpener(self)
-        control_socket = Client('fake_socket_path', opener=mock_opener)
+        control_socket = ControlSocketClient('fake_socket_path', opener=mock_opener)
 
         mock_opener.expect(
             url='http://localhost/metrics-users',
@@ -52,7 +54,7 @@ class TestClass(unittest.TestCase):
 
     def test_remove_metrics_user_success(self):
         mock_opener = MockOpener(self)
-        control_socket = Client('fake_socket_path', opener=mock_opener)
+        control_socket = ControlSocketClient('fake_socket_path', opener=mock_opener)
 
         mock_opener.expect(
             url='http://localhost/metrics-users/juju-metrics-r0',
@@ -68,7 +70,7 @@ class TestClass(unittest.TestCase):
 
     def test_remove_metrics_user_fail(self):
         mock_opener = MockOpener(self)
-        control_socket = Client('fake_socket_path', opener=mock_opener)
+        control_socket = ControlSocketClient('fake_socket_path', opener=mock_opener)
 
         mock_opener.expect(
             url='http://localhost/metrics-users/juju-metrics-r0',
@@ -93,7 +95,7 @@ class TestClass(unittest.TestCase):
 
     def test_connection_error(self):
         mock_opener = MockOpener(self)
-        control_socket = Client('fake_socket_path', opener=mock_opener)
+        control_socket = ControlSocketClient('fake_socket_path', opener=mock_opener)
 
         mock_opener.expect(
             url='http://localhost/metrics-users',
@@ -105,6 +107,18 @@ class TestClass(unittest.TestCase):
 
         with self.assertRaisesRegex(ConnectionError, 'could not connect to socket'):
             control_socket.add_metrics_user('juju-metrics-r0', 'passwd')
+
+    def test_reload_config(self):
+        mock_opener = MockOpener(self)
+        config_reload_socket = ConfigChangeSocketClient('fake_socket_path', opener=mock_opener)
+
+        mock_opener.expect(
+            url='http://localhost/reload',
+            method='GET',
+            body=None,
+            response=None,
+        )
+        config_reload_socket.reload_config()
 
 
 class MockOpener:
@@ -127,14 +141,14 @@ class MockOpener:
         else:
             self.test.assertEqual(request.data.decode('utf-8'), self.body)
 
-        if self.response:
-            return self.response
-        else:
+        if self.error:
             raise self.error
+        else:
+            return self.response
 
 
 class MockResponse:
-    def __init__(self, headers, body):
+    def __init__(self, headers, body=None):
         self.headers = headers
         self.body = body
 
@@ -143,7 +157,7 @@ class MockResponse:
 
 
 class MockHeaders:
-    def __init__(self, content_type, params=None):
+    def __init__(self, content_type=None, params=None):
         self.content_type = content_type
         self.params = params
